Continue module decomposition: extract extension auth template
Move extensionAuthTemplate from templates.js into templates/extension-auth-template.js and keep templates.js exports unchanged.

Preserve extension authorization page behavior while shrinking the main template hotspot.
